feat: specify environment explicitly
BREAKING CHANGE: This removes the env utils that tries to detect
the current environment, and requires the consumer to handle this
itself and specify it during initialization.
