+++ README.md
@@ -6,7 +6,7 @@
 initSentry({
   release: 'ec1c1c',
   buildTimestamp: '2019-04-11T00:06+02:00', // Optional
-  isProd: true/false // Optional
+  environment: "...", // Optional, e.g. prod
   sentryDsn: '...'
 })
 ```
@@ -14,32 +14,12 @@ initSentry({
 ## Features
 
 - Throttling of error messages to prevent flooding Sentry
-- Determining environment for application based on origin
 - Determining if Sentry is enabled for the application
 - Provides default configuration for Sentry initialization
 
 ### Throttling
 
 Evaluate and throttle before sending message to Sentry in case of excessive amount of failures. This will allow for a spike of 4-5 requests, decaying using mean lifetime of 1 minute.
-
-### Determining environment for application based on origin
-
-Eligible environment tags in url (such as 'https://dev.company.xyz') are as follows:
-
-- dev
-- test
-- qa
-- devtest
-- systest
-
-They will be determined based on origin. Currently supporting:
-
-- http://localhost:{PORT} --> local
-- https://{Environment}.company.xyz --> {Environment}
-- https://some-project-{Environment}.company.xyz --> {Environment}
-- https://{Environment}-some-project.company.xyz --> {Environment}
-- https://some-{Environment}-project.company.xyz --> {Environment}
-- https://www.{Environment}.company.xyz --> {Environment}
 
 ### Determining if Sentry is enabled for the application
