--- src/types.ts
@@ -1,6 +1,0 @@
-export interface IRequiredConfiguration {
-  release: string;
-  buildTimestamp?: string;
-  isProd?: boolean;
-  sentryDsn: string;
-}
